requirements.txt
flask==3.0.3
flask-cors==4.0.0
pillow==10.3.0
numpy==1.26.4
transparent-background==1.2.13
torch==2.2.2
awsgi==0.2.14